requirements.txt
botocore==1.29.41
boto3==1.26.41
aws-cdk-lib>=2.0.0
constructs>=10.0.0
aws_lambda_powertools==2.5.0
types-boto3==1.0.2
types-setuptools==65.6.0.2